btrfs-progs: Introduce btrfs-modify tool to modify btrfs internal structures
This patch introduce the basic framework of a new prog, btrfs-modify,
which is designed to modify variant internal btrfs structures and
provide the basis for btrfs recovery test cases.

Btrfs-modify prog uses the command group facility to make it extendable,
the final objective is to replace the less maintained and poorly
documented btrfs-corrupt prog.

Signed-off-by: Qu Wenruo <quwenruo@cn.fujitsu.com>