btrfs-progs: qgroup-verify: Don't treat qgroup difference as error if the fs hasn't initialized a rescan
During test btrfs/166, it's possible to hit a certain case where qgroup
is just enabled but rescan hasn't been kicked in.

Since at qgroup enable time, we mark INCONSISTENT flag, and let later
rescan to clear that flag, if power loss before we kick in rescan, it's
possible we get a qgroup status item with ON|INCONSISTENT but without
RESCAN flag.

And in that case, it will definitely cause difference in qgroup numbers
(as all numbers in qgroup tree is 0).

Fix this false alert by also checking rescan progress from
btrfs_status_item.
And if we find rescan progress is still 0, INCONSISTENT flag set and no
RESCAN flag set, we won't treat it as an error.

Reported-by: Lu Fengqi <lufq.fnst@cn.fujitsu.com>
Signed-off-by: Qu Wenruo <wqu@suse.com>