btrfs-progs: rescue: Add ability to disable quota offline
Provide an offline tool to disable quota.

For kernel which skip_balance doesn't work, there is no way to disable
quota on huge fs with balance, as quota will cause balance to hang for a
long long time for each tree block switch.

So add an offline rescue tool to disable quota.

Reported-by: Dan Merillat <dan.merillat@gmail.com>
Signed-off-by: Qu Wenruo <wqu@suse.com>